+++ src/Dynamics365Configuration/Install-Dynamics365LanguageUpdate.psm1
@@ -94,7 +94,9 @@ function Install-Dynamics365LanguageUpdate {
         }
     } until ( $job.State -eq "Completed" )
 
-    Write-Output "$(Get-Date) Job is complete, output:";
+    $elapsedTime = $( Get-Date ) - $startTime;
+    $elapsedString = "{0:HH:mm:ss}" -f ( [datetime]$elapsedTime.Ticks );
+    Write-Output "$(Get-Date) Elapsed $elapsedString. Job is complete, output:";
     Write-Output ( Receive-Job $job );
     Remove-Job $job;
     $currentProductInstalled = Get-ItemProperty HKLM:\Software\Microsoft\Windows\CurrentVersion\Uninstall\* | Where-Object { $_.PSChildName -eq "{$( $Dynamics365Resources.( $baseResourceName ).identifyingNumber )}" }