+++ src/Dynamics365Configuration/Install-Dynamics365ReportingExtensionsUpdate.psm1
@@ -87,7 +87,9 @@ function Install-Dynamics365ReportingExtensionsUpdate {
                 }
             } until ( $job.State -eq "Completed" )
 
-            Write-Output "$(Get-Date) Job is complete, output:";
+            $elapsedTime = $( Get-Date ) - $startTime;
+            $elapsedString = "{0:HH:mm:ss}" -f ( [datetime]$elapsedTime.Ticks );
+            Write-Output "$(Get-Date) Elapsed $elapsedString. Job is complete, output:";
             Write-Output ( Receive-Job $job );
             Remove-Job $job;
         }