+++ src/Dynamics365Configuration/Install-Dynamics365Server.psm1
@@ -527,7 +527,9 @@ function Install-Dynamics365Server {
                             }
                         } until ( $job.State -eq "Completed" )
 
-                        Write-Output "$(Get-Date) Job is complete, output:";
+                        $elapsedTime = $( Get-Date ) - $startTime;
+                        $elapsedString = "{0:HH:mm:ss}" -f ( [datetime]$elapsedTime.Ticks );
+                        Write-Output "$(Get-Date) Elapsed $elapsedString. Job is complete, output:";
                         Write-Output ( Receive-Job $job );
 
                         Remove-Job $job;